+++ threedigrid/admin/gridresultadmin.py
@@ -13,6 +13,7 @@ from threedigrid.admin.breaches.models import Breaches
 from threedigrid.admin.breaches.timeseries_mixin import (
     BreachesAggregateResultsMixin,
     BreachesResultsMixin,
+    get_breaches_customized_result_mixin,
 )
 from threedigrid.admin.constants import DEFAULT_CHUNK_TIMESERIES
 from threedigrid.admin.gridadmin import GridH5Admin
@@ -20,17 +21,20 @@ from threedigrid.admin.h5py_datasource import H5pyResultGroup
 from threedigrid.admin.h5py_swmr import H5SwmrFile
 from threedigrid.admin.lines.models import Lines
 from threedigrid.admin.lines.timeseries_mixin import (
+    get_lines_customized_result_mixin,
     LinesAggregateResultsMixin,
     LinesResultsMixin,
 )
 from threedigrid.admin.nodes.models import Nodes
 from threedigrid.admin.nodes.timeseries_mixin import (
+    get_nodes_customized_results_mixin,
     get_substance_result_mixin,
     NodesAggregateResultsMixin,
     NodesResultsMixin,
 )
 from threedigrid.admin.pumps.models import Pumps
 from threedigrid.admin.pumps.timeseries_mixin import (
+    get_pumps_customized_result_mixin,
     PumpsAggregateResultsMixin,
     PumpsResultsMixin,
 )
@@ -64,8 +68,8 @@ class GridH5ResultAdmin(GridH5Admin):
         :param file_modus: modus in which to open the files
         """
         self._field_model_dict = defaultdict(list)
-        self._netcdf_file_path = netcdf_file_path
-        super().__init__(h5_file_path, file_modus)
+        self._netcdf_file_path = str(netcdf_file_path)
+        super().__init__(str(h5_file_path), file_modus)
 
         self.result_datasource_class = H5pyResultGroup
 
@@ -610,3 +614,210 @@ class GridH5WaterQualityResultAdmin(GridH5Admin):
     def close(self) -> None:
         super().close()
         self.netcdf_file.close()
+
+
+class CustomizedResultAdmin(GridH5Admin):
+    """Interface for customized 3Di result files
+
+    Customized 3Di result files are result files where users can specify nodes, lines,
+    and pumps of interest, combined with specific output variables and intervals.
+    This interface can be used to extract data from customized result files. It
+    contains functions to extract data from nodes, lines, breaches, and pumps.
+    Additionally, it can be used to extract data from the result area of interest.
+
+        >>> cra = CustomizedResultAdmin(gridadmin_path, customized_results_3di.nc)
+        >>> cra.nodes.id
+        >>> cra.lines.id
+        >>> cra.area1.nodes.subset("1D_ALL").id
+        >>> cra.area1.lines.subset("2D_ALL").s1
+        >>> cra.area1.breaches.id
+        >>> cra.area2.pumps.id
+    """
+
+    def __init__(
+        self,
+        h5_file_path: str,
+        netcdf_file_path: str,
+        file_modus: str = "r",
+        swmr: bool = False,
+    ) -> None:
+        """
+        :param h5_file_path: path to the hdf5 gridadmin file
+        :param netcdf_file_path: path to the water quality result file
+            (usually structure_control_actions_3di.nc)
+        :param file_modus: modus in which to open the files
+        """
+        super().__init__(h5_file_path, file_modus)
+
+        self._netcdf_file_path: str = netcdf_file_path
+        if swmr:
+            self.netcdf_file = H5SwmrFile(netcdf_file_path, file_modus)
+        else:
+            self.netcdf_file = h5py.File(netcdf_file_path, file_modus)
+
+        self.set_timeseries_chunk_size(DEFAULT_CHUNK_TIMESERIES.stop)
+        self.netcdf_keys = self.netcdf_file.keys()
+
+        for key in self.netcdf_keys:
+            regex_match = re.search(r"Mesh\d{1,2}D(Node|Line|Pump)_id_area\d+", key)
+            if regex_match:
+                area_name = regex_match.group().split("_")[-1]
+                if not hasattr(self, area_name):
+                    self.__setattr__(
+                        area_name, _CustomizedAreaResultAdmin(self, area_name)
+                    )
+
+    def close(self) -> None:
+        super().close()
+        self.netcdf_file.close()
+
+    @property
+    def nodes(self):
+        """Build nodes interface if there are nodes present in the result file."""
+        if not hasattr(self, "_nodes"):
+            self._nodes = self._build_nodes_result_group("")
+        return self._nodes
+
+    @property
+    def lines(self):
+        """Build lines interface if there are lines present in the result file."""
+        if not hasattr(self, "_lines"):
+            self._lines = self._build_lines_result_group("")
+        return self._lines
+
+    @property
+    def breaches(self):
+        if not hasattr(self, "_breaches"):
+            if not self.has_breaches:
+                self._breaches = None
+            else:
+                self._breaches = self._build_breaches_result_group("")
+
+        if self._breaches is None:
+            logger.info("Threedimodel has no breaches")
+
+        return self._breaches
+
+    @property
+    def pumps(self):
+        if not hasattr(self, "_pumps"):
+            if not self.has_pumpstations:
+                self._pumps = None
+            else:
+                self._pumps = self._build_pumps_result_group("")
+
+        if self._pumps is None:
+            logger.info("Threedimodel has no pumps")
+
+        return self._pumps
+
+    def _build_nodes_result_group(self, area: str) -> Optional[Nodes]:
+        size = self.result_group_size(f"nMesh2D_nodes{area}", f"nMesh1D_nodes{area}")
+        if size == 0:
+            return None
+
+        return Nodes(
+            H5pyResultGroup(self.h5py_file, "nodes", self.netcdf_file),
+            **dict(
+                self._grid_kwargs,
+                **{"mixin": get_nodes_customized_results_mixin(self.netcdf_keys, area)},
+            ),
+        )
+
+    def _build_lines_result_group(self, area: str) -> Optional[Lines]:
+        size = self.result_group_size(f"nMesh2D_lines{area}", f"nMesh1D_lines{area}")
+        if size == 0:
+            return None
+
+        return Lines(
+            H5pyResultGroup(self.h5py_file, "lines", self.netcdf_file),
+            **dict(
+                self._grid_kwargs,
+                **{
+                    "mixin": get_lines_customized_result_mixin(
+                        self.netcdf_keys,
+                        area,
+                    )
+                },
+            ),
+        )
+
+    def _build_breaches_result_group(self, area: str) -> Breaches:
+        return Breaches(
+            H5pyResultGroup(self.h5py_file, "breaches", self.netcdf_file),
+            **dict(
+                self._grid_kwargs,
+                **{
+                    "mixin": get_breaches_customized_result_mixin(
+                        self.netcdf_keys,
+                        area,
+                    )
+                },
+            ),
+        )
+
+    def _build_pumps_result_group(self, area: str) -> Pumps:
+        return Pumps(
+            H5pyResultGroup(self.h5py_file, "pumps", self.netcdf_file),
+            **dict(
+                self._grid_kwargs,
+                **{
+                    "mixin": get_pumps_customized_result_mixin(
+                        self.netcdf_keys,
+                        area,
+                    ),
+                },
+            ),
+        )
+
+    def result_group_size(self, field_name_2d: str, field_name_1d) -> np.ndarray:
+        two_d_size = self.netcdf_file.get(field_name_2d, np.array([])).size
+        one_d_size = self.netcdf_file.get(field_name_1d, np.array([])).size
+        return two_d_size + one_d_size
+
+    def set_timeseries_chunk_size(self, new_chunk_size):
+        """overwrite the default chunk size for timeseries queries"""
+        _chunk_size = int(new_chunk_size)
+        if _chunk_size < 1:
+            raise ValueError("Chunk size must be greater than 0")
+        self._timeseries_chunk_size = slice(0, _chunk_size)
+        logger.info("New chunk for timeseries size has been set to %d", new_chunk_size)
+        self._grid_kwargs.update({"timeseries_chunk_size": self._timeseries_chunk_size})
+
+
+class _CustomizedAreaResultAdmin:
+    """Nested class for customized 3Di result files to interact with specific areas"""
+
+    def __init__(
+        self, customized_result_admin: CustomizedResultAdmin, area_name: str
+    ) -> None:
+        self.cra = customized_result_admin
+        self.area_name = area_name
+        self._timeseries_chunk_size = customized_result_admin._timeseries_chunk_size
+        self._grid_kwargs = customized_result_admin._grid_kwargs
+
+    @property
+    def nodes(self):
+        """Build nodes interface if there are nodes present in the result file."""
+        if not hasattr(self, "_nodes"):
+            self._nodes = self.cra._build_nodes_result_group(f"_{self.area_name}")
+        return self._nodes
+
+    @property
+    def lines(self):
+        """Build lines interface if there are lines present in the result file."""
+        if not hasattr(self, "_lines"):
+            self._lines = self.cra._build_lines_result_group(f"_{self.area_name}")
+        return self._lines
+
+    @property
+    def breaches(self):
+        if not hasattr(self, "_breaches"):
+            self._breaches = self.cra._build_breaches_result_group(f"_{self.area_name}")
+        return self._breaches
+
+    @property
+    def pumps(self):
+        if not hasattr(self, "_pumps"):
+            self._pumps = self.cra._build_pumps_result_group(f"_{self.area_name}")
+        return self._pumps