+++ threedigrid/admin/lines/timeseries_mixin.py
@@ -1,5 +1,8 @@
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
 
+from typing import List
+
+from threedigrid.admin.constants import NO_DATA_VALUE
 from threedigrid.orm.base.options import ModelMeta
 from threedigrid.orm.base.timeseries_mixin import AggregateResultMixin, ResultMixin
 
@@ -69,3 +72,63 @@ class LinesAggregateResultsMixin(AggregateResultMixin):
         :param kwargs:
         """
         super().__init__(**kwargs)
+
+
+def get_lines_customized_result_mixin(fields: List[str], area: str):
+    def construct_line_customized_base_composite_fields(fields: List[str], area: str):
+        """ID is added as a composite field as it is not equal to the grid ids"""
+        composite_fields = {}
+        for field in fields:
+            for key, values in BASE_COMPOSITE_FIELDS.items():
+                if field in values:
+                    if key not in composite_fields:
+                        composite_fields[key] = [field]
+                    else:
+                        # Sort just to be sure 2D is before 1D
+                        composite_fields[key].insert(0, field)
+                        composite_fields[key].sort(reverse=True)
+
+        if "_mesh_id" in composite_fields:
+            for i, v in enumerate(composite_fields["_mesh_id"]):
+                composite_fields["_mesh_id"][i] = f"{v}{area}"
+
+        composite_fields["id"] = []
+        if "Mesh2DLine_id" in fields:
+            composite_fields["id"].insert(0, "Mesh2DLine_id")
+        if "Mesh1DLine_id" in fields:
+            composite_fields["id"].append("Mesh1DLine_id")
+
+        composite_fields["kcu"] = []
+        if "Mesh2DLine_type" in fields:
+            composite_fields["kcu"].insert(0, "Mesh2DLine_type")
+        if "Mesh1DLine_type" in fields:
+            composite_fields["kcu"].append("Mesh1DLine_type")
+
+        return composite_fields
+
+    def construct_line_customized_base_subset_fields(fields: List[str]):
+        subset_fields = {}
+        for key, value in BASE_SUBSET_FIELDS.items():
+            subset = list(value.keys())[0]
+            if value[subset] in fields:
+                subset_fields[key] = {subset: value[subset]}
+
+        return subset_fields
+
+    composites = construct_line_customized_base_composite_fields(fields, area)
+    subsets = construct_line_customized_base_subset_fields(fields)
+
+    class LinesCustomizedResultsMixin(LinesResultsMixin):
+        class Meta:
+            field_attrs = ["units", "long_name", "standard_name"]
+
+            composite_fields = composites
+            subset_fields = subsets
+            lookup_fields = ("_mesh_id", "id")
+            is_customized_mixin = True
+            composite_field_insert_values = {"kcu": int(NO_DATA_VALUE)}
+
+        def __init__(self, **kwargs):
+            super().__init__(**kwargs)
+
+    return LinesCustomizedResultsMixin