+++ threedigrid/admin/pumps/timeseries_mixin.py
@@ -1,7 +1,5 @@
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
-"""
-"""
-
+from typing import List
 
 from threedigrid.orm.base.options import ModelMeta
 from threedigrid.orm.base.timeseries_mixin import AggregateResultMixin, ResultMixin
@@ -47,3 +45,23 @@ class PumpsAggregateResultsMixin(AggregateResultMixin):
         :param kwargs:
         """
         super().__init__(**kwargs)
+
+
+def get_pumps_customized_result_mixin(fields: List[str], area: str):
+    if f"Mesh1DPump_id{area}" not in fields:
+        return
+
+    composites = {}
+    composites["id"] = ["Mesh1DPump_id"]
+    composites["_mesh_id"] = [f"Mesh1DPump_id{area}"]
+    if "Mesh1D_q_pump" in fields:
+        composites["q_pump"] = ["Mesh1D_q_pump"]
+
+    class PumpsCustomizedResultsMixin(ResultMixin):
+        class Meta:
+            field_attrs = ["units", "long_name", "standard_name"]
+            composite_fields = composites
+            lookup_fields = ("_mesh_id", "id")
+            is_customized_mixins = True
+
+    return PumpsCustomizedResultsMixin